// src/collector/helpers/interfaces/settings.ts
export interface ICollectorSettings {
    /** Specifies if block enumerators (question numbers) should be displayed. */
    enumerators: boolean;

    /** Specifies if the page indicators should be displayed (only available in paginated mode). */
    pages: boolean;

    /** Specifies if the progressbar should be displayed. */
    progressbar: boolean;
}